refactor(users): simplify who to follow structure
