+++ src/components/organisms/FeedRightSidebar/FeedRightSidebar.tsx
@@ -11,7 +11,7 @@
 import { Pencil, UsersRound } from 'lucide-react';
 import { Container } from '@/atoms/Container/Container';
 import { FeedSection } from '@/molecules/FeedSection/FeedSection';
-import { WhoToFollowSidebar } from '@/organisms/WhoToFollow/WhoToFollowSidebar/WhoToFollowSidebar';
+import { WhoToFollowSidebar } from '@/organisms/WhoToFollowSidebar/WhoToFollowSidebar';
 import { ActiveUsers } from '../ActiveUsers/ActiveUsers';
 import { FeedbackCard } from '../FeedbackCard/FeedbackCard';
 import { HotTags } from '../HotTags/HotTags';
